app/requirements.txt
Django>=4.2
django-ckeditor>=6.5.1
django-debug-toolbar>=4.1.0
mixer>=7.2.2
Pillow>=9.5.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests>=2.31.0
six>=1.16.0
sorl-thumbnail>=12.9.0
django-environ>=0.9
